+++ src/main/java/gift/repository/product/ProductRepository.java
@@ -1,9 +1,8 @@
 package gift.repository.product;
 
-import gift.domain.Product;
+import gift.model.product.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
-
 
 }
